Minor: combine dataframe tests
